Реорганизовал h_battery.cpp
