+++ server/entities/func_recharge.cpp
@@ -20,35 +20,7 @@
 
 */
 
-#include "extdll.h"
-#include "util.h"
-#include "cbase.h"
-#include "saverestore.h"
-#include "skill.h"
-#include "gamerules.h"
-#include "player.h"
-
-class CRecharge : public CBaseToggle
-{
-	DECLARE_CLASS( CRecharge, CBaseToggle );
-public:
-	void Spawn( );
-	void Precache( void );
-	void Off(void);
-	void Recharge(void);
-	void KeyValue( KeyValueData *pkvd );
-	void Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value );
-	virtual int ObjectCaps( void ) { return (BaseClass :: ObjectCaps() | FCAP_CONTINUOUS_USE) & ~FCAP_ACROSS_TRANSITION; }
-	virtual STATE GetState( void );
-
-	DECLARE_DATADESC();
-
-	float m_flNextCharge; 
-	int m_iReactivate ; // DeathMatch Delay until reactvated
-	int m_iJuice;
-	int m_iOn;	// 0 = off, 1 = startup, 2 = going
-	float m_flSoundTime;
-};
+#include "func_recharge.h"
 
 BEGIN_DATADESC( CRecharge )
 	DEFINE_FIELD( m_flNextCharge, FIELD_TIME ),
