+++ server/entities/func_recharge.h
@@ -1,0 +1,53 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	
+*	This product contains software technology licensed from Id 
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+/*
+
+===== h_battery.cpp ========================================================
+
+  battery-related code
+
+*/
+
+#pragma once
+
+#include "extdll.h"
+#include "util.h"
+#include "cbase.h"
+#include "saverestore.h"
+#include "skill.h"
+#include "gamerules.h"
+#include "player.h"
+
+class CRecharge : public CBaseToggle
+{
+	DECLARE_CLASS( CRecharge, CBaseToggle );
+public:
+	void Spawn( );
+	void Precache( void );
+	void Off(void);
+	void Recharge(void);
+	void KeyValue( KeyValueData *pkvd );
+	void Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value );
+	virtual int ObjectCaps( void ) { return (BaseClass :: ObjectCaps() | FCAP_CONTINUOUS_USE) & ~FCAP_ACROSS_TRANSITION; }
+	virtual STATE GetState( void );
+
+	DECLARE_DATADESC();
+
+	float m_flNextCharge; 
+	int m_iReactivate ; // DeathMatch Delay until reactvated
+	int m_iJuice;
+	int m_iOn;	// 0 = off, 1 = startup, 2 = going
+	float m_flSoundTime;
+};
